Most basic version finished
Last thing that was added was a cleaned up if and else if statements in which it would know which button the user is pressing.
If this button was an operator, it would take the first number, store it in the second number place,
then look for the "first number" again.
While this works, it is not the best has subtraction and division functions num1 and num2 had to be switched.
Things to fix:
Javascript:
-the result stays after another number is entered. This new number goes on the end of the result
-User validation
-If user does 2 + 2, then hits + again, this should display the result of 2+2 while having the + operator pressed.
-Add keyboard support
html/css:
The design needs to be updated.